+++ NEWS.md
@@ -9,6 +9,11 @@
 - Implemented modules in the Shiny app.
 - Added a registry for optional arguments for specific geoms. List them using `geom_args()`.
 - Shiny app now displays optional geom arguments along with explanatory comments.
+- Add arg `percent` to `hd_opts()` to show `%` in tooltip and y-axis if `TRUE`.
+- Add arg `xtick` to `hd_opts` to select column for custom x‑axis tick labels
+  when these differ from the numeric x values. Note that Highcharts uses
+  zero-based indexing, while R uses one-based indexing.
+
 
 # highdir 0.3.0